requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
numpy
python-multipart
requests
pandas
gunicorn